Checks inputs to codiv()

Checks inputs to codiv()

Usage

check_inputs(
  Host_tree,
  Symbiont_tree,
  Host_to_Symbiont_df,
  min_hosts,
  min_symbiont_tips,
  span_fraction,
  permutations
)

Arguments

Host_tree

A binary phylogenetic tree of hosts

Symbiont_tree

A binary phylogenetic tree of symbionts

Host_to_Symbiont_df

A data.frame() that pairs all of the tips in the Symbiont_tree with the tip in the Host_tree it was isolated from.

min_hosts

The minimum number hosts for a node to be evaluated (must be greater than 3)

min_symbiont_tips

The minimum number symbiont tips for a node to be evaluated (recommend 7 or more)

span_fraction

The fraction of the Symbiont_tree included in the codiv scan

permutations

The number of permutations (recommend at least 99)

Value

Tests to make sure the inputs are acceptable

Examples

Host_tree <- ape::rtree(10)
Symbiont_tree <- ape::rtree(100)
Host_to_Symbiont_df <- data.frame("Host" = rep(Host_tree$tip.label, 10), 
                                    "Symbiont" = Symbiont_tree$tip.label)
check_inputs(Host_tree, Symbiont_tree, Host_to_Symbiont_df, 3, 7, 0.25, 99)
#> `Host_tree` and `Symbiont_tree` Inputs must be binary phylogentic trees
#> `Symbiont_tree` contains tips that are not in`Host_to_Symbiont_df`